music player menu by components
